Fix bug in pushing projections in nested CTEs
There was a bug in CTE projection and filter pushdown involving chained CTEs like in the following query:

WITH cte1 AS (SELECT * FROM ORDERS WHERE orderkey < 1000),
cte5 AS (SELECT orderkey FROM cte1 WHERE totalprice < 100000)
SELECT * FROM cte5;

The problem was that when applying projection pushdown, the rewrite was not recursing into the source of the CTEProducer, which resulted in CTEConsumer and CTEProducer to have mismatched columns.
This surfaced as "expr not found" errors.

The fix adds recursive step during the rewrite when encountering a CTEProducer node.